Consolidate all outlier detection tests into single file
- Merged tests from test_pca_outliers.py (11 PCA reconstruction tests)
- Merged tests from test_outlier_indices_validation.py (15 index validation tests)
- Removed duplicate test files to maintain clean repository structure
- All 54 tests passing successfully
